Fix ktlint: move AppUpdateResult to separate file
Ktlint requires that files containing a single class be named after that
class. Moved AppUpdateResult sealed class from AppUpdateManager.kt to its
own file AppUpdateResult.kt.
