+++ composeApp/src/commonMain/kotlin/com/meneses/budgethunter/commons/platform/AppUpdateManager.kt
@@ -1,11 +1,4 @@
 package com.meneses.budgethunter.commons.platform
 
-sealed class AppUpdateResult {
-    data object NoUpdateAvailable : AppUpdateResult()
-    data object UpdateInProgress : AppUpdateResult()
-    data class UpdateAvailable(val startUpdate: () -> Unit) : AppUpdateResult()
-    data object UpdateFailed : AppUpdateResult()
-}
-
 // Platform-specific implementations are provided in androidMain and iosMain
 // Use IAppUpdateManager interface for dependency injection
