+++ composeApp/src/commonMain/kotlin/com/meneses/budgethunter/commons/platform/AppUpdateResult.kt
@@ -1,0 +1,8 @@
+package com.meneses.budgethunter.commons.platform
+
+sealed class AppUpdateResult {
+    data object NoUpdateAvailable : AppUpdateResult()
+    data object UpdateInProgress : AppUpdateResult()
+    data class UpdateAvailable(val startUpdate: () -> Unit) : AppUpdateResult()
+    data object UpdateFailed : AppUpdateResult()
+}
